Add workflows submodule and update documentation.
- CLI argument parsing is in scripts while the actual pipeline
(stringing together functions in the package to produce output) is in
workflows.
- This will update the doc pages automatically based on changes to the
code, and will allow users to access workflows in pure Python.
