+++ .gitignore
@@ -1,3 +1,5 @@
+docs/generated/
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
