+++ docs/api.rst
@@ -1,5 +1,165 @@
 API
 ===
 
-:mod:`rapidtide.io_funcs`: Functions
+
+:mod:`rapidtide.workflows`: Rapidtide workflows
 --------------------------------------------------
+.. automodule:: rapidtide.workflows
+    :no-members:
+    :no-inherited-members:
+
+.. autosummary:: rapidtide.workflows
+    :toctree: generated/
+    :template: function.rst
+
+    rapidtide.workflows.rapidtide2
+
+.. currentmodule:: rapidtide
+
+.. _calibration_ref:
+
+
+:mod:`rapidtide.correlate`: Correlation functions
+--------------------------------------------------
+.. automodule:: rapidtide.correlate
+    :no-members:
+    :no-inherited-members:
+
+.. autosummary:: rapidtide.correlate
+    :toctree: generated/
+    :template: function.rst
+
+    rapidtide.correlate.autocorrcheck
+    rapidtide.correlate.quickcorr
+
+.. currentmodule:: rapidtide
+
+.. _calibration_ref:
+
+
+:mod:`rapidtide.filter`: Filters
+--------------------------------------------------
+.. automodule:: rapidtide.filter
+    :no-members:
+    :no-inherited-members:
+
+.. autosummary:: rapidtide.filter
+    :toctree: generated/
+    :template: function.rst
+
+    rapidtide.filter.padvec
+    rapidtide.filter.unpadvec
+
+.. currentmodule:: rapidtide
+
+.. _calibration_ref:
+
+
+:mod:`rapidtide.fit`: Fitting functions
+--------------------------------------------------
+.. automodule:: rapidtide.fit
+    :no-members:
+    :no-inherited-members:
+
+.. autosummary:: rapidtide.fit
+    :toctree: generated/
+    :template: function.rst
+
+    rapidtide.fit.gaussresidualssk
+    rapidtide.fit.gaussskresiduals
+
+.. currentmodule:: rapidtide
+
+.. _calibration_ref:
+
+
+:mod:`rapidtide.io`: Input/output functions
+--------------------------------------------------
+.. automodule:: rapidtide.io
+    :no-members:
+    :no-inherited-members:
+
+.. autosummary:: rapidtide.io
+    :toctree: generated/
+    :template: function.rst
+
+    rapidtide.io.checkifparfile
+    rapidtide.io.readbidssidecar
+
+.. currentmodule:: rapidtide
+
+.. _calibration_ref:
+
+
+:mod:`rapidtide.miscmath`: Miscellaneous math functions
+-------------------------------------------------------
+.. automodule:: rapidtide.miscmath
+    :no-members:
+    :no-inherited-members:
+
+.. autosummary:: rapidtide.miscmath
+    :toctree: generated/
+    :template: function.rst
+
+    rapidtide.miscmath.phase
+    rapidtide.miscmath.polarfft
+
+.. currentmodule:: rapidtide
+
+.. _calibration_ref:
+
+
+:mod:`rapidtide.resample`: Resampling functions
+--------------------------------------------------
+.. automodule:: rapidtide.resample
+    :no-members:
+    :no-inherited-members:
+
+.. autosummary:: rapidtide.resample
+    :toctree: generated/
+    :template: function.rst
+
+    rapidtide.resample.congrid
+
+    :toctree: generated/
+    :template: class.rst
+
+    rapidtide.resample.fastresampler
+
+.. currentmodule:: rapidtide
+
+.. _calibration_ref:
+
+
+:mod:`rapidtide.stats`: Statistical functions
+--------------------------------------------------
+.. automodule:: rapidtide.stats
+    :no-members:
+    :no-inherited-members:
+
+.. autosummary:: rapidtide.stats
+    :toctree: generated/
+    :template: function.rst
+
+    rapidtide.stats.printthresholds
+
+.. currentmodule:: rapidtide
+
+.. _calibration_ref:
+
+
+:mod:`rapidtide.util`: Utility functions
+--------------------------------------------------
+.. automodule:: rapidtide.util
+    :no-members:
+    :no-inherited-members:
+
+.. autosummary:: rapidtide.util
+    :toctree: generated/
+    :template: function.rst
+
+    rapidtide.util.logmem
+
+.. currentmodule:: rapidtide
+
+.. _calibration_ref:
