+++ rapidtide/workflows/__init__.py
@@ -1,0 +1,8 @@
+# emacs: -*- mode: python-mode; py-indent-offset: 4; tab-width: 4; -*-
+# ex: set sts=4 ts=4 sw=4 et:
+
+# Overrides submodules with their functions.
+from .rapidtide2 import rapidtide2
+
+
+__all__ = ['rapidtide2']
